Added the --delay=# or #-# (varying) option
The main loop passes a kwargs dict full of keyword args to every server thread
spawned, which is used to collect and transmit static and mutable arguments to
the servers.  If a varying delay=#-# range is specified, a thread is spawned to
regularly mutate the delay.value in that range.